optlib: provide flags to print messages from a regex entry
It's useful to be able to print messages to stderr within regex tables,
to warn about unexpected conditions or even exit ctags. So this
adds a `{fatal="some message"}` and `{warning="some other message"}`
flags for regex options. I saw no reason to limit it to multi-table
ones only, so instead it applies to all regex types.