unit: accept input_[0-9].* and input_[0-9]_*.* as input file names
Java language expects a class and the file where the class is defined
have a same name. On the othre hand, the language doesn't allow to use
`-` in a class name.

This means that we cannot write a test input file as a valid Java program.
This change solves this situation.

Using `[', `]' pair instead of `{', `}' is suggested by @b4n.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>