main: don't check role enablement in makeSimpleRefTag
This change is suggested by @hadrielk in #1855.

In #1855, we remove the code for chekcing kind enablement from
makeSimpleTag. This change does the same about role enablement.

Instead of removing the code for checking role enablement, this change
adds the code for checking `name' argument as makeSimpleTag does.

Current universal-ctags doesn't provide the way to enable or disable a
role. However, keeping the code consistency is important for
maintenance.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>