tex: Fix infinite loop parsing an identifier at EOF
This was introduced in 5d5afe64ca5517c50897164f72db174428961e9f which
started accepting as an identifier character any byte with a value that
cast as `unsigned char` is greater than or equal to 0x80.  The common
value for the `EOF` constant being -1, which cast to `unsigned char`
gives a value of 0xff, hence leading to accepting `EOF` as an
identifier character and then an infinite loop.

Fix it on both sides:

* Don't perform the cast, as it's unneeded in this case: `c` is an
  `int` holding the value of an `unsigned char` or `EOF` already, so
  anything equal to or above 0 is a legitimate input character.
* Guard against `EOF` explicitly in the loop.

Fixes #1896.