main,Tcl: provide the way to skip autoFQ tag emission
Consider a parser that utilizes autoFQ tag emission feature PARTIALLY.
The parser expects the main part emits autoFQ tags for some of normal tags.
However, the parser doesn't want the main part to emit autoFQ tags for the rest
of normal tags because the parser makes FQ tags for them by itself.
Tcl parser is a such parser.

This change provides a parser a way to notify "don't make autoFQ tag
for this tag" to the main part.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>