PHP: use anonGenerate function
anonGenerate function defined in main part is for generating names
for anonymous tags. Unlike the current particular implementation
in php.c, the name generated by anonGenerate function is unique in
a tags file. The unique names may help a client tool implement
interesting features.

In addition to using the function, this change marks tags for
anonymous classes XTAG_ANONYMOUS. So a user can enable or disable
the emission of tags for anonymous classes with --extras='-{anonymous}'
option.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>