HTML: lift up the code to get an attribute when parsing a HTML tag
The original code gets an attribute when the parser finds an
interesting HTML tag. This approach is efficient because the parser
doesn't have to get the attribute for non-interesting HTML tag.

However if parser has interests in variety HTML tags, the code
becomes redundant.

This change lifts up the code to get an attribute. With this change,
the parser gets attributes always; it doesn't concern HTML tags
surrounding the attributes. Not efficient but adding a new interesting
HTML tag becomes easier.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>