main: accept whiespace in tagEntryInfo even in e-ctags output mode
The original code rejects printing a tagEntryInfo that includes whiespace or tab chars
in e-ctags output mode.

This is reported as a bug in #2014 by @elecalion.
I myself also reported in #1969.

What we should reject is just tab char.

NOTE: This char doesn't fix the original bug perfectly.  If --sort=no
is given, e-ctags output mode doesn't reject a tagEntryInfo including
'\t'. The code for rejecting depends on wrong assumption that the
output stream (to tags file) is rewindable. This is not true if the
output is connected to stdout directly. The options combination, "-o
-" and "--sort=no" breaks my assumption.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>