main: don't use stream rewinding to implement e-ctags output
To remove a tag entry including tab char in e-ctags output mode,
ctags rewinded the output stream after printing th entry.

We find this implementation doesn't work well if both "-o -" and
"--sort=no" options are given at one time.  In this case, the output
stream is not rewindable.

This chnage introduces hasTagEntryTabChar() and checks whther a
tag entry is printable or not before prnting.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>