main: increment the number of written tags only when a tag is not rejected
This fixes a bug.

Our interanl sort routine refers the number of written tags.
The original code increments the number even if a tag to be written is
rejected because it includes a tab char. As the result, ctags emits
more tags than expected when "--sort=yes -o -" optoins are given.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>