C++: record template type parameters to detect the end of template perfix
If we know Foo is a name of type, it becomes eadier to detect whether
">>" in "Foo>>" is a shift operator or the end marker of th template
prefix.

In this change, a strlist instance is introduced to cxx_parser_template.c
for tracking template type parameters in a template prefix.
When cxxParserParseTemplateAngleBracketsInternal() needs to detect
whether ">>" is an operator or end markers, the strlist instance gives
a hint to the function.

Making a tag for a template type parameter is next challenge.  If the
parameters are tagged, corkQueue can be used instead of the strlist
instance.

Clsoe #2060.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>