main: fix a bug not calling inputStart() method of a subparser
If a subparser uses addLanguageCallbackRegex, ctags didn't call
a inputStart() subparser method. This is a bug.

ctags should not call inputStart() of a subparser if
METHOD_NOT_CRAFTED is set to the method field of the subparser.

There is a critical typo to extract the it for METHOD_NOT_CRAFTED from
a flag.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>