Cobol: introduce a parser specific parameter named "format" to switch cobol dialects about source code format
@b4n introduced CobolFree and CobolVariable parsers when he rewrote
Cobol parser.  Ideally, such dialects can be handled in a parser. So
we can keep the output of --list-languages small. However, there is no
reliable way to recognize the format dialect used in a cobol file.

Instead of introducing unreliable automatic dialect guessing code,
this change introduces a parser specific parameter named "format".

    [jet@living]~/var/ctags% ./ctags --list-params=Cobol
    #NAME   DESCRIPTION
    format  source code format: [fixed], free, or variable
    [jet@living]~/var/ctags% ./ctags --param-cobol:format=fixed
    ./ctags --param-cobol:format=fixed
    ctags: No files specified. Try "ctags --help".
    [jet@living]~/var/ctags% ./ctags --param-cobol:format=free
    ./ctags --param-cobol:format=free
    ctags: No files specified. Try "ctags --help".
    [jet@living]~/var/ctags% ./ctags --param-cobol:format=variable
    ./ctags --param-cobol:format=variable
    ctags: No files specified. Try "ctags --help".
    [jet@living]~/var/ctags% ./ctags --param-cobol:format=xxx
    ./ctags --param-cobol:format=xxx
    ctags: Warning: Cobol: an unexpcted value is given as an argument to "format" parameter: xxx
    ctags: Warning: Cobol: reset "format" to "fixed"
    ctags: No files specified. Try "ctags --help".

A user can specify a format manually instead of choosing a parser.
--list-languages prints just one Cobol parser.

    [jet@living]~/var/ctags% ./ctags --list-languages | grep Cobol
    Cobol

This change does more things.
CobolFree and CobolVariable parsers are still there.  Further more,
this change adds CobolFixed parser.  However, "invisible" parser flags
of them are set true. As the result, they are not appeared in --list-languages
output.

So Geany can specify a parser from the 3 parsers explicitly.  This is
a hack. When we add an interface to give a value to the parser
specific parameter to libctags, we can remove the 3 parsers
specialized to each dialect.