main,json-writer: allow the writer to disable "fixed" fields
Close #2065.

The tags format must include name, input, and pattern fields.
When running ctags, disabling them are not allowed. To implement
this rule, "fixed" fields were introduced to field.c.

However, this rule is applicable only to the ctags writer. The
other writer like json-writer doesn't need to follow the rule.

The fixed field is implemented in field.c. Now, it moves to writer-ctags.c.
As the result, json-writer can print the result with disabling
"name", "input", and/or "pattern".

Signed-off-by: Masatake YAMATO <yamato@redhat.com>