Sample client app using ctags as a library plus all the related changes in ctags
This patch demonstrates how we use ctags in Geany as a library. The
patch tries to be minimalistic to demonstrate only the things we
need in ctags so it just directly replaces the ctags code with what
we need and the generated "ctags" binary is the demo app. For this
to be mergeable to ctags, there have to be some ifdefs specifying
whether ctags should be compiled as a library and mini-geany has
to be a separate binary.

In addition, @masatake did:
* fix a memleak about mio allocation.
* build both ctags and mini-geany
* include writer_p.h in parser.c (for writerRescanFailed)
* unify setTagWriter and setCustomTagWriter
* ...