main: rename main() in main.c to ctags_main_cli()
This is a part of commits for introducing libctags.a.
To add main.o to libctags.a, main.c should not have main()
function. As the temporary solution, this commit changes the
name of main() to ctags_cli_main().

This commit introduces cmd.c which will not be part of libctags.a.
The real main() function is put to cmd.c.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>