main: extend --totals option to print parsers specific statistics
initStats and printStats methods are added to parserDefinition struct.

If a parser fills the methods, the parser specific statistics are
printed with --totals=extra is given.

The main part calls the initStats method of the current parser if
--totals=extra is given. The parser can prepare privaite data
structure for gathering its statistics.

The printStats method of the current parser is called at the end of
ctags execution. As the name shown, the method is for printing the

Signed-off-by: Masatake YAMATO <yamato@redhat.com>