main: allow specifying alternative parser list
Allow applications which use ctags as a library provide their own
list of parsers. This list has to be declared as a define called
EXTERNAL_PARSER_LIST inside an external header. ctags will use the
EXTERNAL_PARSER_LIST_FILE define to detect the presence of such a
header and it should contain path to the header. When not defined,
internal parsers are used.