Add Geany parsers which are missing in ctags
In addition to these parsers we also have a simple markdown parser
which however just uses readLine() interface and isn't token-based
so it's about equivalent to the regex-based parser in ctags (which
generates more kinds and is simpler so better to use that one). Plus
there are some more parsers inside c.c which however needs much more
work to merge with ctags.

I don't really know much about these parsers - some of them are for
really obscure languages like abaqus but they have been in Geany for
ages and seem to work.

This commit uses the parsers from pull request

https://github.com/geany/geany/pull/2132

in which were some minor modifications so these parsers compile with
upstream ctags.