NSIS: new parser imported from Geany
This pull request adds one of the extra parsers we have in Geany except
those defined in c.c and except the markdown parser which is readLine()
based and not really superior to the regex-based parser ctags uses
(we should probably use the ctags parser in Geany instead).

The added a parser is for Nsis (https://en.wikipedia.org/wiki/Nullsoft_Scriptable_Install_System).

The original version is written for Geany by Enrico Tröger <enrico.troeger@uvena.de>:
commit 61f89d5c259c2538ceff815e8f2e5226343e89ff
Author: Enrico Tröger <enrico.troeger@uvena.de>
Date:   Tue Apr 21 20:52:19 2009 +0000

    Add a trivial symbol parser for NSIS files.

    git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@3716 ea778897-0a13-0410-b9d1-a72fbfd435f5

This log message is edited by @masatake.