Use kind definitions specific for the given languages
In Geany we used to modify all parsers so all of them generated the
same set of kinds which we then process internally the same way.
Now it is possible in Geany to redefine the ctags kinds to whatever
we need so these hacks are not necessary.

This patch modifies the kind names and kind letters so they better
match the real use in the affected parsers.