JSON: limit the depth of recursion by tracking balancing of bracket pairs
Close #2107 reported by @hanwen.
Related to #1682.

To avoid stack overflow, this change provides the way to terminate
the parsing if recursion CAN be too deep.

The recursion is very related to square brackets and curly brackets
in input stream.
Instead of limit the depth of function recursion itself, this change
tracks the depth of brackets. Here I assume deeply nested brackets
increases the recursion depth of function calls.

readTokenFull is changed to return EOF token when it detects too
deeply nested (>= 512) brackets in the current input stream. The
EOF token may terminate the parsing GENTLY.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>