main: introduce a function putting a placeholder tag in the corkQueue
Consider following input written in imaginary programming language X:

 A {
   B;
 }

B have A as its parent scope.

With using the cork API, you can make tags with the following step:

1. make the tag for A, and get the cork index (index_A) for A, and
2. make the tag for B filling scopeIndex field with index_A.

In the most of all case, these steps work well.

However, there is an exception.  The language X has following odd
character; the kind of A can be decided after parsing B. For such
language character, above steps don't work.

You cannot make the tag for A before making B because the kind for A
cannot be decided till parsing B. However, you cannot make the tag for
B before making the tag for A because index_A is needed to make the
tag for B. Mutual dependencies are there.

This change introduces makePlaceholder function. This function extends
the corkAPI to work well even if the mutual dependencies are there.

makePlaceholder() allows to make a temporary tag (placeholder).
Reconsider the language X. When the parser finds A, the parser can
call makePlaceholder(A). The parser can make a tag for A without
specifying its kind. It returns the corkIndex for the placeholder. The
parser can use it index_A when parsing and making tag for B. The
parser gets the cork index (index_B) for B.

After parsing B, now the parser knows the real kind for A.  The parser
can convert the placeholder tag to the real tag for A by patching the
placeholder tag; the parser refills the kindIndex field of A with the
real kind.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>