XML: improve namespace handling
The code detecting namespace with no prefix had a bug.

    <svg xmlns="http://www.w3.org/2000/svg">
    ...

This is an example of "namespace with no prefix".
XML parser detects it to make an anonymous tag having
namespace kind.

    <svg>
    ...

The old code detected a XML element having no "xmlns"
as "namespace with no prefix", and emitted an anonymous tag.
This is wrong. This change fixes the wrong behavior.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>