[SELF INCOMPATIBLE] SVG: convert to a subparser of XML parser
SVG parser captured IDs in <tag id=ID> by its own tagXpathTableTable.
However, id=ID is nothing special to SVG.
The id attributes are common in any XML files.
On the other hand, XML parser has the ability to capture d=ID.

This change converts SVG parser to a subparser of XML parser.
id=ID in a svg file is captured as usual but it is done by XML parser.
It means the language fields of the tags are changed from SVG to XML.

SELF INCOMPATIBLE change: Now, the id kind in SVG parser is gone.
XML's one is used instead.

In addition, "def" kind is introduced to SVG parser. It is used only
for tagging id="ID" **in <def/>**. id=ID attributes in <defs/> are
captured TWICE; first time by XML parser, and the second time by SVG
parser.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>