[SELF INCOMPATIBLE] Glade: convert the parser to a subparser of XML parser
Glade parser captured IDs in <tag id=ID> by its own tagXpathTableTable.
However, id=ID is nothing special to Glade.
The id attributes are common in any XML files.
On the other hand, XML parser has the ability to capture d=ID.

This change converts Glade parser to a subparser of XML parser.
id=ID in a glade file is captured as usual but it is done by XML parser.
It means the language fields of the tags are changed from Glade to XML.

SELF INCOMPATIBLE change: Now, the id kind in Glade parser is gone.
XML's one is used instead.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>