main,refactor: call xmlSetGenericErrorFunc from functions calling xmlLineNumbersDefault and xmlParseMemory
xmlSetGenericErrorFunc is called in two places: xmlParseMIO and findXMLTagsFull.

xmlParseMIO also calls xmlLineNumbersDefault, and xmlParseMemory.
From the view of maintenance, it is better that the three functions is
called in a function.

In other hand, xmlParseMIO calls only
xmlSetGenericErrorFunc. xmlParseMIO calls the rest two functions
indirectly. makeXMLDoc calls the two. This change moves the line
calling xmlSetGenericErrorFunc to makeXMLDoc.

In the future, the parts of makeXMLDoc and xmlParseMIO should be
unified.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>