SQL: introduce new field to token to distinguish reserved words from keywords
As reported in #2215 by @akemrir, there is case that a keyword can be
used for a name for table and/or field.

The keywords cannot be used for a name are called reserved words.
For the fixing the original bug, and for future enhancement,
this change introduces reserved_word field to the token structure.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>