main: Suppress MinGW warnings
Cast `off_t` to `long long` and use "%lld" for the format string.

The size of `off_t` is 32-bit by default on Win32 (VC and MinGW),
however it becomes 64-bit when `_FILE_OFFSET_BITS` is defined to 64 on
MinGW-w64 (both 32-bit and 64-bit). It's not so easy to set the format
string properly. Using `long long` is the easiest way.