lregex: fix lang_set to avoid core dump
Markdown regex based parser used `guest` flag.
This fix the bug of guest flag implementation.

Markdown parser runs a guest parser for the area surrounded with
```...```. If the language for the area is specified explicitly
like the following example, ctags works fine:

	```C
	int main(void) { return 0; }
	```

"C" is specified in the example.

However, if the language is not specified explicitly like the following
example, a core dump may occur because referring a wrongly set value:

	 ```
	 int main(void) { return 0; }
	 ```

Even if the core dump doesn't occur on a platform, valgrind reports
a wrong memory access.

(Based on #2261, @masatake wrote this commit log.)