C++: recognize "class" just after '>' as a keyword leading a type parameter
In the following input, to detect Container as a type parameter,
class before Container must be recognized as a class keyword.

	template<template<class...> class Container, class Elem>
	auto foo1(const Container<Elem> & p_container)
	{
	    return Container<Elem>{};
	}

The original logic expects class appears only after '<' or ','.
So the "class" just after '>' is not recognized as a keyword
leading a type parameter.

This change extends the original logic to accept "class" after
">" as the keyword.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>