docs,man: Commit generated man files into docs/man/
Fix the issue in https://github.com/universal-ctags/ctags/pull/2213#issuecomment-536358592.

Replace the placeholders like `@CTAGS_NAME_EXECUTABLE@` with `ctags` and
so on, so that the man pages can be shown correctly on the Web site.

The files under docs/man/ need to be updated manually:

```
$ make -C man QUICK=1 update-docs
```

So a checker for that is added to Travis CI.

Now the docs/man/ files are not symlinks, so they can be correctly
handled on Windows. A workaround on AppVeyor is not needed.