buildsys: Fix that packcc may fail on MSYS2
When shadow builds are used on MSYS2, packcc fails to read *.peg file.
It seems that MSYS's make uses `\` for directory separator on some
conditions. In such situation, a command `./packcc ..\peg\varlink.peg`
is executed, and `\` is considered as an escape character. Then the
command is interpreted as `./packcc ..pegvarlink.peg`, and it fails.