fuzz: remove unneeded backslash passed to --kinds-all and --fields
In a shell script, * is not expanded surrounded by " and ".
Putting \ as prefix for * doesn't make sense.

In addition, replace --*-kinds option with --kinds-all option.  The
feature using * for representing "all languages" was removed ago.  The
order of words in the option is also obsoleted; use --kinds-<LANG>
order instead of --<LANG>-kinds.

Locally tested with "make fuzz".

Signed-off-by: Masatake YAMATO <yamato@redhat.com>