main: use Assert macro to detect wrong parser names
If the name of a built-in parser includes an unacceptable char,
it is a programming bug, not an runtime error.

The checking for `invisible' field is for old CTagsSelfTestParser.
It didn't have neither parser nor parser2 method. However, now
it has parser method. Therefore, there is no need for checking
`invisible' field.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>