lregex: fix meaningless type mismatching between unsigned int and unsigned long
Partially close #2369.

A pointer to unsigned int variable is passed to functions expecting
a pointer to unsigned long variable.

This didn't cause a trouble because the sizes of the types are the
same on x86_64, the most popular platform. However, on s390x, they
are different, and cause a trouble as reported in #2369.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>