Markdown: don't recognized a hairline as a GFM (sub)section
The parser recognized a line filled with white spaces before a
hairline marker (===== or ----) as a section or subsection.
The wrong recognition has no impact on tag emission. However,
it causes warning messages.

To suppress the warning messages, introduce a table for just ignoring
a line filled with white spaces.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>