Units,dts: remove unnecessary Z field specification from args.ctags
The case dts-fq-without-scope-field.d is for testing the conflicting
between the feature generating automatic full-qualified tags and the
command line interface disabling scope field.

Z option originally enabled in args.ctags is for prefixing scope:
key to the scope field. It doesn't control enabling/disabling
the field. So specifying Z is not needed in the test case.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>