enumstr.sh: support token type using char literal
Use char literal as the type of token in a popular technique in
parsers.

python.c:

    typedef enum eTokenType {
	    /* 0..255 are the byte's value */
	    TOKEN_EOF = 256,
	    TOKEN_UNDEFINED,
	    TOKEN_INDENT,
	    TOKEN_KEYWORD,
	    TOKEN_OPERATOR,
	    TOKEN_IDENTIFIER,
	    TOKEN_STRING,
	    TOKEN_WHITESPACE,
    } tokenType;

...in tokenizer:

	case '/':
	{
		int d = getcFromInputFile ();
		vStringPut (token->string, c);
		if (d != '=')
		{
			ungetcToInputFile (d);
>>>>>			token->type = c;
		}

enumstr.sh can generate a function converting
a enum element to a C string. However, it it
not enough for debugging purpose. It could not
convert the types using a byte value between 0..255.

This change extends the script to support type
implicitly defined with the byte values.

If newly introduced --use-lower-bits-as-is option
is given, the script emits a function extended for
supporting the byte values.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>