[INCOMPATIBLE]Tex: introduce "xinput" kind as the replacement for "include" kind
foo in \include{foo} doesn't define foo.
The statement is for including foo, a file already existing on file system.
The statement just refers it. Therefore, it should be captured as a
reference tag.

This change introduces new kind "xinput", and captures foo in
\include{foo} with "included" role of "xinput" kind.

In addition, the change captures bar in \input{bar} with
"input" role of "xinput" kind.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>