Go: don't check whether kinds is enabled or no at the parser side
Close #2430.

returned from getEnryFromCorkIndex().

Nowadays, Go parser utilizes corkIndex various ways to fill fields
of tag entries.

makeTagFull functio of Go parser checks whether the kinds of Go
are enabled or disabled; if a given kind is disabled, the
function doesn't make a tag for the kind. As the result the function
returns CORK_NIL. So CORK_NIL is not suitable for filling the fields.

The main part of ctags has generic layer for doing the same as
the parser; the main part also checks whether a kind assigned to
a tagEntryInfo is enabled or not. For the optimization purpose,
early checking in a parser is good. However, it makes the code
for using cork indexes in parser sides complicated; a parser
must check the value returned from getEnryFromCorkIndex().

Signed-off-by: Masatake YAMATO <yamato@redhat.com>